Add some more models.
BareBear.vox is the old Bear.vox before I added grass, a flower, and a mushroom. Castle.vox is derived closely from a default model in MagicaVoxel. Grin.vox is a requested user icon to match the style of Bear.vox, and it helps when checking the behavior of very dark (with some very bright) voxels.